working on the app

+++ backend/src/resolvers/Mutation.js
@@ -38,7 +38,14 @@ const Mutations = {
 
     async deleteItem(parent, args, ctx, info) {
         const where = { id: args.id };
-        const item = await ctx.db.query.item({ where }, `{ id title }`);
+        const item = await ctx.db.query.item({ where }, `{ id title user {id}}`);
+
+        const ownsItem = item.user.id === ctx.request.userId;       
+        const hasPermission = ctx.request.user.permissions.some(permission => ["ADMIN", "ITEMDELETE"].includes(permission));
+
+        if(!hasPermission || !ownsItem) {
+            throw new Error("You do not have permissions to delete this item");
+        }
 
         return await ctx.db.mutation.deleteItem({ where }, info);
     },
